requirements.txt
colorama==0.4.6
Pillow==10.2.0
pytesseract==0.3.10
Requests==2.31.0
tqdm==4.66.2
